Календарь (#72)

+++ kanbanboard/Forms/MainForm.Table.cs
@@ -1,10 +1,10 @@
-﻿using System;
+﻿using kanbanboard.Classes;
+using kanbanboard.Controls;
+using System;
 using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
 using System.Windows.Forms;
-using kanbanboard.Classes;
-using kanbanboard.Controls;
 
 namespace kanbanboard.Forms
 {
@@ -26,12 +26,29 @@ namespace kanbanboard.Forms
                 for (var row = 1; row < TableLayoutPanel.RowStyles.Count; row++)
                 {
                     if (TableLayoutPanel.GetControlFromPosition(column, row) != null)
+                    {
+
                         ticket.Add(new Dictionary<string, string>
                         {
-                            {"Title", ((TicketPanel)TableLayoutPanel.GetControlFromPosition(column, row)).Title.Text},
-                            {"Ticket", ((TicketPanel)TableLayoutPanel.GetControlFromPosition(column, row)).Ticket.Text},
-                            {"People", ((TicketPanel)TableLayoutPanel.GetControlFromPosition(column, row)).People.Text}
+                            {
+                                "Title",
+                                ((TicketPanel) TableLayoutPanel.GetControlFromPosition(column, row)).Title.Text},
+                            {
+                                "Ticket",
+                                ((TicketPanel) TableLayoutPanel.GetControlFromPosition(column, row)).Ticket.Text
+                            },
+                            {
+                                "People",
+                                ((TicketPanel) TableLayoutPanel.GetControlFromPosition(column, row)).People.Text
+                            },
+                            {
+                                "Deadline",
+                                $"{((TicketPanel)TableLayoutPanel.GetControlFromPosition(column, row)).Deadline.Day}." +
+                                $"{((TicketPanel)TableLayoutPanel.GetControlFromPosition(column, row)).Deadline.Month}." +
+                                $"{((TicketPanel)TableLayoutPanel.GetControlFromPosition(column, row)).Deadline.Year}"
+                            },
                         });
+                    }
                 }
 
                 if ((TitlePanel)TableLayoutPanel.GetControlFromPosition(column, 0) != null)
@@ -50,7 +67,7 @@ namespace kanbanboard.Forms
         }
 
         // Загрузить данные в tablelayoutpanel
-        private void TableFromFirebase(string selectedProject)
+        private async void TableFromFirebase(string selectedProject)
         {
             LoadPanel.BringToFront();
 
@@ -71,14 +88,16 @@ namespace kanbanboard.Forms
                     foreach (var tasks in titles.Value)
                     {
                         var ticket = new TicketPanel();
-                        try {  
-                            ticket.Title.Text = tasks["Title"] ?? ""; 
-                            ticket.Ticket.Text = tasks["Ticket"] ?? ""; 
+                        try
+                        {
+                            ticket.Title.Text = tasks["Title"] ?? "";
+                            ticket.Ticket.Text = tasks["Ticket"] ?? "";
                             ticket.People.Text = tasks["People"] ?? "";
+                            ticket.Deadline = DateTime.Parse(tasks["Deadline"]);
                         }
                         catch { }
 
-                        AddControlToPanel(ticket.Title.Text, ticket.Ticket.Text, ticket.People.Text, column, row++);
+                        AddControlToPanel(ticket.Title.Text, ticket.Ticket.Text, ticket.People.Text, ticket.Deadline, column, row++);
                     }
                     column++;
                     row = 1;
@@ -87,8 +106,12 @@ namespace kanbanboard.Forms
 
             ResizeTable();
 
+            // Показываем всех работников на странице календаря
+            await ShowEmployees().ConfigureAwait(false);
+
             if (BasicContentPanel.Controls.GetChildIndex(PanelWithTable) == 1) PanelWithTable.BringToFront();
             else if (BasicContentPanel.Controls.GetChildIndex(MessengerPanel) == 1) MessengerPanel.BringToFront();
+            else if (BasicContentPanel.Controls.GetChildIndex(CalendarPanel) == 1) CalendarPanel.BringToFront();
         }
 
         // События на кнопки
@@ -168,7 +191,7 @@ namespace kanbanboard.Forms
                     for (var row = 1; row <= TableLayoutPanel.RowStyles.Count; row++)
                     {
                         if (!(TableLayoutPanel.GetControlFromPosition(column, row) is null)) continue;
-                        AddControlToPanel("Заголовок", "Описание", "Разработчики", column, row);
+                        AddControlToPanel("Заголовок", "Описание", "Разработчики", DateTime.Now, column, row);
                         break;
                     }
                 };
@@ -239,7 +262,7 @@ namespace kanbanboard.Forms
 
             if (TableLayoutPanel.ColumnStyles.Count >= 1 && column == 0) column = TableLayoutPanel.ColumnStyles.Count;
 
-            AddControlToPanel("Заголовок", "Описание", "Разработчики", column, 1);
+            AddControlToPanel("Заголовок", "Описание", "Разработчики", DateTime.Now, column, 1);
             AddTitleToPanel("Это тайтл", column);
         }
 
@@ -270,12 +293,13 @@ namespace kanbanboard.Forms
         }
 
         // Добавить тикет в таблицу
-        private void AddControlToPanel(string title, string ticket, string people, int column, int row)
+        private void AddControlToPanel(string title, string ticket, string people, DateTime date, int column, int row)
         {
             var control = new TicketPanel();
             control.Title.Text = title;
             control.Ticket.Text = ticket;
             control.People.Text = people;
+            control.Deadline = date;
 
             // Добавляем события на кнопки, если пользователь соответствующей роли
             if (_user.Role is Roles.Admin || _user.Role is Roles.Manager) SetEventsOnTicket(control);
@@ -324,7 +348,8 @@ namespace kanbanboard.Forms
         {
             try
             {
-                foreach (ColumnStyle column in TableLayoutPanel.ColumnStyles) {
+                foreach (ColumnStyle column in TableLayoutPanel.ColumnStyles)
+                {
                     column.SizeType = SizeType.Percent;
                     column.Width = 25;
                 }
@@ -333,13 +358,15 @@ namespace kanbanboard.Forms
 
                 TableLayoutPanel.RowStyles[0].SizeType = SizeType.Absolute;
                 TableLayoutPanel.RowStyles[0].Height = 25;
-                
-                foreach (var row in TableLayoutPanel.RowStyles.Cast<RowStyle>().ToList().Skip(1)) {
+
+                foreach (var row in TableLayoutPanel.RowStyles.Cast<RowStyle>().ToList().Skip(1))
+                {
                     row.SizeType = SizeType.Absolute;
                     row.Height = 150;
                 }
 
-                TableLayoutPanel.Controls.OfType<TicketPanel>().ToList().ForEach(x => {
+                TableLayoutPanel.Controls.OfType<TicketPanel>().ToList().ForEach(x =>
+                {
                     if (TableLayoutPanel.GetCellPosition(x).Row != 0)
                         x.Height = TableLayoutPanel.Height / TableLayoutPanel.RowCount;
                 });
@@ -361,8 +388,14 @@ namespace kanbanboard.Forms
         // Удалить выбранный проект
         private async void TrashButton_Click(object sender, EventArgs e)
         {
+            if (MessageBox.Show("Удалить доску из твоего профиля? (действие отменить невозможно)", "Удаление доски", MessageBoxButtons.YesNo, MessageBoxIcon.Question) != DialogResult.Yes)
+                return;
+
             if (await _user.DeleteProject(ListBoxOfProjectNames.SelectedItem?.ToString()) == "OK")
+            {
                 SetUserProjectNames();
+                TableLayoutPanel.Controls.Clear();
+            }
         }
     }
 }
